+++ api/projects-api/package.json
@@ -15,6 +15,7 @@
     "@hono/zod-validator": "catalog:",
     "@incmix-api/utils": "workspace:^",
     "@incmix/utils": "catalog:",
+    "bullmq": "catalog:",
     "hono": "catalog:",
     "kysely": "catalog:",
     "kysely-ctl": "catalog:",
