+++ api/projects-api/src/routes/index.ts
@@ -3,7 +3,9 @@ import healthcheckRoutes from "@/routes/healthcheck"
 import type { HonoApp } from "@/types"
 import type { OpenAPIHono } from "@hono/zod-openapi"
 import projectRoutes from "./projects"
+import tasksRoutes from "./tasks"
 export const routes = (app: OpenAPIHono<HonoApp>) => {
   app.route(`${BASE_PATH}/healthcheck`, healthcheckRoutes)
   app.route(BASE_PATH, projectRoutes)
+  app.route(`${BASE_PATH}/tasks`, tasksRoutes)
 }
