--- api/tasks-api/.dockerignore
@@ -1,5 +1,0 @@
-node_modules
-.turbo
-.git
-dist
-.env
