--- api/tasks-api/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@incmix-api/tasks-api",
-  "type": "module",
-  "scripts": {
-    "dev": "tsx watch src/index.ts",
-    "build": "tsc && tsc-alias",
-    "start": "node ./dist/index.js",
-    "test": "vitest --run",
-    "test:e2e": "playwright test",
-    "clean": "rm -rf dist"
-  },
-  "dependencies": {
-
-    "@hono/zod-openapi": "catalog:",
-    "@hono/zod-validator": "catalog:",
-    "@incmix-api/utils": "workspace:^",
-    "@incmix/utils": "catalog:",
-    "bullmq": "catalog:",
-    "hono": "catalog:",
-    "kysely": "catalog:",
-    "kysely-ctl": "catalog:",
-    "luxon": "catalog:",
-    "nanoid": "catalog:",
-    "pg": "catalog:",
-    "zod": "catalog:"
-  },
-  "devDependencies": {
-    "@incmix-api/config": "workspace:^",
-    "@playwright/test": "catalog:",
-    "@types/node": "catalog:",
-    "@types/pg": "catalog:",
-    "tsc-alias": "catalog:",
-    "tsx": "catalog:",
-    "typescript": "catalog:",
-    "vite-tsconfig-paths": "catalog:",
-    "vitest": "catalog:"
-  }
-}
