--- api/tasks-api/src/routes/healthcheck/openapi.ts
@@ -1,20 +1,0 @@
-import { HealthCheckSchema } from "@/routes/healthcheck/types"
-import { createRoute } from "@hono/zod-openapi"
-
-export const healthCheck = createRoute({
-  path: "/",
-  method: "get",
-  tags: ["Healthcheck"],
-  security: [{ cookieAuth: [] }],
-  summary: "Check Service Health",
-  responses: {
-    200: {
-      content: {
-        "application/json": {
-          schema: HealthCheckSchema,
-        },
-      },
-      description: "Returns Service Status",
-    },
-  },
-})
