--- api/tasks-api/src/routes/types.ts
@@ -1,9 +1,0 @@
-import { z } from "@hono/zod-openapi"
-
-export const ResponseSchema = z
-  .object({
-    message: z.string().openapi({
-      example: "Successful",
-    }),
-  })
-  .openapi("Response")
