--- api/tasks-api/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "extends": "@incmix-api/config/hono.json",
-  "compilerOptions": {
-    "baseUrl": "./",
-    "outDir": "./dist",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  },
-  "tsc-alias": {
-    "resolveFullPaths": true
-  },
-  "include": ["src/**/*.ts"],
-  "exclude": [
-    "node_modules",
-    "e2e",
-    "test",
-    "**/test/**",
-    "**/e2e/**",
-    "playwright.config.ts"
-  ]
-}
